change format locale from mainjs to stringjs and more test
